Move ILikeAny, LikeAny in snowflake, use initial implementation for base
