+++ sqlglot/tokens.py
@@ -528,7 +528,6 @@ class Tokenizer(metaclass=_Tokenizer):
         "HAVING": TokenType.HAVING,
         "IF": TokenType.IF,
         "ILIKE": TokenType.ILIKE,
-        "ILIKE ANY": TokenType.ILIKE_ANY,
         "IGNORE NULLS": TokenType.IGNORE_NULLS,
         "IN": TokenType.IN,
         "INDEX": TokenType.INDEX,
@@ -545,7 +544,6 @@ class Tokenizer(metaclass=_Tokenizer):
         "LEADING": TokenType.LEADING,
         "LEFT": TokenType.LEFT,
         "LIKE": TokenType.LIKE,
-        "LIKE ANY": TokenType.LIKE_ANY,
         "LIMIT": TokenType.LIMIT,
         "LOAD DATA": TokenType.LOAD_DATA,
         "LOCAL": TokenType.LOCAL,
